requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
docopt==0.6.2
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
requests==2.26.0
urllib3==1.26.7
yarg==0.1.9
